Fix bug with new commanders for old slots
When creating a new commander, it was inserted into a map of slot number
to commander. However, std::unordered_map::emplace only inserts the
element if it does not already exist in the map. This was causing newly
created commanders to simply restart the already existing commander
which had its exit_ flag set to true. Then, Commander::Run() was called
on a commander that had been killed, causing a fault.

The fix is to first remove the tracked commander for the slot, if one
exists.