Fix bug with receiving too many P2B messages
If a majority of servers in the system is three servers but five are
online, it is possible to receive P2Bs after a decision has been
reached. The commander for a log slot is destructed after a decision has
been reached for the slot. Receiving a P2B once the commander is
destructed would previously cause the assertion to fail, causing a crash
in the system. This fix updates the assert to simply check if a
commander exists for the slot, only asking the commander to handle the
P2B if it exists.